optimizing `fused_grouped_topk` SYCL kernel for MoE expert routing
