+++ csrc/moe/moe_ops.h
@@ -56,18 +56,6 @@ std::tuple<torch::Tensor, torch::Tensor> fused_grouped_topk(
     const double routed_scaling_factor,
     const c10::optional<torch::Tensor>& bias);
 
-std::tuple<torch::Tensor, torch::Tensor> grouped_topk_multi_group(
-    const torch::Tensor& hidden_states,
-    const torch::Tensor& gating_output,
-    const int64_t n_topk,
-    const bool renormalize,
-    const int64_t n_expert_group,
-    const int64_t n_topk_group,
-    const c10::string_view scoring_func,
-    const double routed_scaling_factor,
-    const c10::optional<torch::Tensor>& bias);
-
-
 
 void topk_softmax(
     torch::Tensor& topk_weights,
