+++ tests/register_ops.py
@@ -473,24 +473,6 @@ def fused_grouped_topk(
                                                routed_scaling_factor,
                                                e_score_correction_bias)
 
-def grouped_topk_multi_group(
-    hidden_states: torch.Tensor,
-    gating_output: torch.Tensor,
-    topk: int,
-    renormalize: bool,
-    num_expert_group: int,
-    topk_group: int,
-    scoring_func: str = "softmax",
-    routed_scaling_factor: float = 1.0,
-    e_score_correction_bias: Optional[torch.Tensor] = None,
-):
-    return torch.ops._moe_C.grouped_topk_multi_group(hidden_states, gating_output,
-                                               topk, renormalize,
-                                               num_expert_group, topk_group,
-                                               scoring_func,
-                                               routed_scaling_factor,
-                                               e_score_correction_bias)
-
 
 def topk_softmax(topk_weights: torch.Tensor, topk_ids: torch.Tensor,
                  token_expert_indices: torch.Tensor,
