+++ tests/test_grouped_topk.py
@@ -3,7 +3,7 @@ import pytest
 import torch
 
 from tests.ops.grouped_topk_op import (fused_grouped_topk,
-                                       fused_grouped_topk_sycl, grouped_topk,grouped_topk_multi_group)
+                                       fused_grouped_topk_sycl, grouped_topk)
 from tests.utils import seed_everything
 
 #override pytest parameters when enable mini pytest
@@ -18,14 +18,14 @@ MINI_PYTEST_PARAMS = {
 }
 
 
-@pytest.mark.parametrize("n_token", [1, 33, 64, 50000,100000])
+@pytest.mark.parametrize("n_token", [64, 50000,100000])
 @pytest.mark.parametrize("n_hidden", [1024, 2048])
-@pytest.mark.parametrize("n_expert", [256])
+@pytest.mark.parametrize("n_expert", [128, 256])
 @pytest.mark.parametrize("topk", [8])
-@pytest.mark.parametrize("renormalize", [False,True])
+@pytest.mark.parametrize("renormalize", [True, False])
 @pytest.mark.parametrize("num_expert_group", [8])
 @pytest.mark.parametrize("topk_group", [4])
-@pytest.mark.parametrize("scoring_func", ["sigmoid",'softmax'])
+@pytest.mark.parametrize("scoring_func", ["sigmoid","softmax"])
 @pytest.mark.parametrize("routed_scaling_factor", [1.0, 2.5])
 @pytest.mark.parametrize("dtype",
                          [torch.float16, torch.bfloat16, torch.float32])
@@ -36,10 +36,10 @@ def test_grouped_topk(n_token: int, n_hidden: int, n_expert: int, topk: int,
     seed_everything(0)
     hidden_states = torch.randn((n_token, n_hidden), dtype=dtype, device="xpu")
     gating_output = torch.randn((n_token, n_expert), dtype=dtype, device="xpu")
+    
     e_score_correction_bias = torch.randn((n_expert, ),
                                           dtype=dtype,
                                           device="xpu")
-
     baseline_topk_weights, baseline_topk_ids = grouped_topk(
         hidden_states=hidden_states,
         gating_output=gating_output,
@@ -50,7 +50,6 @@ def test_grouped_topk(n_token: int, n_hidden: int, n_expert: int, topk: int,
         scoring_func=scoring_func,
         routed_scaling_factor=routed_scaling_factor,
         e_score_correction_bias=e_score_correction_bias)
-
     test_topk_weights, test_topk_ids = fused_grouped_topk(
         hidden_states=hidden_states,
         gating_output=gating_output,
@@ -72,40 +71,22 @@ def test_grouped_topk(n_token: int, n_hidden: int, n_expert: int, topk: int,
         scoring_func=scoring_func,
         routed_scaling_factor=routed_scaling_factor,
         e_score_correction_bias=e_score_correction_bias)
-    test_topk_weights_multi_group, test_topk_ids_multi_group = grouped_topk_multi_group(
-        hidden_states=hidden_states,
-        gating_output=gating_output,
-        topk=topk,
-        renormalize=renormalize,
-        num_expert_group=num_expert_group,
-        topk_group=topk_group,
-        scoring_func=scoring_func,
-        routed_scaling_factor=routed_scaling_factor,
-        e_score_correction_bias=e_score_correction_bias)
 
     if renormalize:
-        # torch.testing.assert_close(baseline_topk_weights,
-        #                            test_topk_weights,
-        #                            atol=2e-2,
-        #                            rtol=0)
-        # torch.testing.assert_close(baseline_topk_weights,
-        #                            test_topk_weights_sycl,
-        #                            atol=2e-2,
-        #                            rtol=0)
         torch.testing.assert_close(baseline_topk_weights,
-                                   test_topk_weights_multi_group,
+                                   test_topk_weights,
+                                   atol=2e-2,
+                                   rtol=0)
+        torch.testing.assert_close(baseline_topk_weights,
+                                   test_topk_weights_sycl,
                                    atol=2e-2,
                                    rtol=0)
 
-    # torch.testing.assert_close(baseline_topk_ids,
-    #                            test_topk_ids,
-    #                            atol=0,
-    #                            rtol=0)
-    # torch.testing.assert_close(baseline_topk_ids,
-    #                            test_topk_ids_sycl,
-    #                            atol=0,
-    #                            rtol=0)
     torch.testing.assert_close(baseline_topk_ids,
-                                   test_topk_ids_multi_group,
-                                   atol=0,
-                                   rtol=0)
+                               test_topk_ids,
+                               atol=0,
+                               rtol=0)
+    torch.testing.assert_close(baseline_topk_ids,
+                               test_topk_ids_sycl,
+                               atol=0,
+                               rtol=0)
